+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Ampache API:
   * Method `get_similar`
   * Type `album_artist` to the method `get_indexes` (but without support for the args `filter`, `add`, `update`)
+  * Fields `time`, `albumcount`, and `songcount` to the `artist` and `album` type results
 - Files playlist tab: Tooltip showing the file path or stream URL
 - Subsonic: Empty implementation for the method `getNowPlaying`
   [#1079](https://github.com/owncloud/music/pull/1079) @NattyNarwhal